Update to the latest snapshot from upstream.
This brings in a few bugfixes and also changes the build system. The
update to the build system means we can drop the gettext patch (although
it may need to make a comeback later because I don't think the new build
system produces .mo files anymore) and simplifies the %install section.

Fixes #8
Fixes #13